LPS-62913 move classes
