+++ modules/apps/portlet-configuration/portlet-configuration-toolbar-contributor-locator-api/src/main/java/com/liferay/portlet/configuration/toolbar/contributor/locator/StrutsPortletToolbarContributorLocator.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.portlet.toolbar.contributor.locator;
+package com.liferay.portlet.configuration.toolbar.contributor.locator;
 
 import com.liferay.portal.kernel.module.framework.ModuleServiceLifecycle;
 import com.liferay.portal.kernel.portlet.toolbar.contributor.locator.PortletToolbarContributorLocator;
@@ -23,10 +23,42 @@ import org.osgi.service.component.annotations.Deactivate;
 import org.osgi.service.component.annotations.Reference;
 
 /**
+ * Provides an implementation of {@link PortletToolbarContributorLocator} for
+ * portlets using Struts as MVC pattern, allowing them to have a different
+ * {@link
+ * com.liferay.portal.kernel.portlet.toolbar.contributor.PortletToolbarContributor}
+ * for different struts actions.
+ *
+ * <p>
+ * PortletToolbarContributor implementations must be registered in the OSGI
+ * Registry using the following properties:
+ * </p>
+ *
+ * <ul>
+ * <li>
+ * &quot;javax.portlet.name&quot; the ID of the portlet whose portlet toolbar to
+ * extend.
+ * </li>
+ * <li>
+ * &quot;struts.action&quot; this property is optional. If this property is not
+ * present, the portlet toolbar is always extended. If it contains a value
+ * (e.g., <code>/blogs/view_entry</code>) the portlet toolbar is extended only
+ * for that specific struts action. If the value is &quot;-&quot; the portlet
+ * toolbar is extended when there is no <code>strutsAction</code> specified in
+ * the request (typically when rendering the first view of the portlet).
+ * </li>
+ * </ul>
+ *
+ * <p>
+ * A single PortletToolbarContributor implementation can be used for different
+ * portlets and struts actions by including multiple times with the same
+ * properties.
+ * </p>
+ *
  * @author Sergio González
  */
 @Component(immediate = true, service = PortletToolbarContributorLocator.class)
-public class MVCCommandPortletToolbarContributorLocator
+public class StrutsPortletToolbarContributorLocator
 	extends BasePortletToolbarContributorLocator {
 
 	@Activate
@@ -43,12 +75,12 @@ public class MVCCommandPortletToolbarContributorLocator
 
 	@Override
 	protected String getParameterName() {
-		return "mvcRenderCommandName";
+		return "struts_action";
 	}
 
 	@Override
 	protected String getPropertyName() {
-		return "mvc.render.command.name";
+		return "struts.action";
 	}
 
 	@Reference(target = ModuleServiceLifecycle.PORTAL_INITIALIZED, unbind = "-")
